requirements.txt
# Commission AI Assistant Requirements - Updated for Windows Compatibility
# First upgrade pip and setuptools
pip>=21.0
setuptools>=60.0
wheel>=0.37.0

# Core dependencies
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0

# AI & ML libraries - Compatible versions
requests>=2.28.0
numpy>=1.21.0
pandas>=1.5.0

# Sentence transformers and dependencies
torch>=1.13.0
transformers>=4.21.0
sentence-transformers>=2.2.0

# ChromaDB - Compatible version
chromadb>=0.4.0

# Ollama client
ollama>=0.1.0

# Data processing
openpyxl>=3.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0

# PDF processing (optional)
# pdfplumber>=0.7.0

# Utilities
python-multipart>=0.0.5
jinja2>=3.1.0
python-dotenv>=0.19.0

# Development tools (optional)
pytest>=7.0.0

# Windows specific (optional)
# pywin32>=306
